refactor: Remove debug-config and all logging from active codebase
Move debug-config.ts to legacy and remove all console logging from active components. StarField and WASM loader now operate silently with no debug output. Legacy components retain debug functionality for backwards compatibility.
