+++ legacy/lib/performance/debug-config.ts
